requirements.txt
# API Server Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.1

# Optional dependencies for enhanced functionality
python-multipart==0.0.6
python-json-logger==2.0.7

# Agent Router Dependencies
pyyaml==6.0.1
prometheus-client==0.19.0

# RAG Dependencies
qdrant-client==1.7.0
sentence-transformers==2.2.2
torch==2.1.0
tiktoken==0.5.1

# Memory (Mem0) Dependencies
mem0ai==0.0.30
redis==5.0.1
numpy==1.24.3

# MCP Dependencies
mcp==0.9.0
pydantic-settings==2.1.0

# OpenTelemetry Dependencies
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-jaeger==1.21.0

# JWT for authentication
pyjwt==2.8.0
cryptography==41.0.7

# Graph Database (Neo4j)
neo4j==5.15.0

# NLP for NER
spacy==3.7.2

# Data Collection Pipeline Dependencies
rq==1.15.1
requests==2.31.0
PyGithub==2.1.1
notion-client==2.2.1
slack-sdk==3.23.0
scikit-learn==1.3.2
schedule==1.2.0
